Evaluate accuracy of a forecast or model

Summarise the performance of the model using accuracy measures. Accuracy measures can be computed directly from models as the one-step-ahead fitted residuals are available. When evaluating accuracy on forecasts, you will need to provide a complete dataset that includes the future data and data used to train the model.

accuracy(object, ...)

# S3 method for mdl_df
accuracy(object, measures = point_accuracy_measures,
  ...)

# S3 method for fbl_ts
accuracy(object, data,
  measures = point_accuracy_measures, ..., by = NULL)

Arguments

object
A model or forecast object

...
Additional arguments to be passed to measures that use it.

measures
A list of accuracy measure functions to compute (such as point_accuracy_measures, interval_accuracy_measures, or distribution_accuracy_measures)

data
A dataset containing the complete model dataset (both training and test data). The training portion of the data will be used in the computation of some accuracy measures, and the test data is used to compute the forecast errors.

by
Variables over which the accuracy is computed (useful for computing across forecast horizons in cross-validation). If by is NULL, groups will be chosen automatically from the key structure.

See also

Examples

library(fable)

#> Error: package ‘fabletools’ required by ‘fable’ could not be found

library(tsibble)
library(tsibbledata)
library(dplyr)

#> 
#> Attaching package: ‘dplyr’

#> The following object is masked from ‘package:tsibble’:
#> 
#>     id

#> The following object is masked from ‘package:testthat’:
#> 
#>     matches

#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag

#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union

fit <- aus_production %>%
  filter(Quarter < yearquarter("2006 Q1")) %>%
  model(ets = ETS(log(Beer) ~ error("M") + trend("Ad") + season("A")))

#> Error in ETS(log(Beer) ~ error("M") + trend("Ad") + season("A")): could not find function "ETS"

# In-sample training accuracy does not require extra data provided.
accuracy(fit)

#> Error in accuracy(fit): object 'fit' not found

# Out-of-sample forecast accuracy requires the future values to compare with.
# All available future data will be used, and a warning will be given if some
# data for the forecast window is unavailable.
fc <- fit %>%
  forecast(h = "5 years")

#> Error in eval(lhs, parent, parent): object 'fit' not found

fc %>%
  accuracy(aus_production)

#> Error in eval(lhs, parent, parent): object 'fc' not found

# It is also possible to compute interval and distributional measures of
# accuracy for models and forecasts which give forecast distributions.
fc %>%
  accuracy(
    aus_production,
    measures = list(interval_accuracy_measures, distribution_accuracy_measures)
  )

#> Error in eval(lhs, parent, parent): object 'fc' not found