Directional accuracy measures

A collection of accuracy measures based on the accuracy of the prediction's direction (say, increasing or decreasing).

MDA(.resid, .actual, na.rm = TRUE, reward = 1, penalty = 0, ...)

MDV(.resid, .actual, na.rm = TRUE, ...)

MDPV(.resid, .actual, na.rm = TRUE, ...)

directional_accuracy_measures

Format

An object of class list of length 3.

Arguments

.resid

A vector of residuals from either the training (model accuracy) or test (forecast accuracy) data.

.actual

A vector of responses matching the fitted values (for forecast accuracy, new_data must be provided).

na.rm

Remove the missing values before calculating the accuracy measure

reward, penalty

The weights given to correct and incorrect predicted directions.

...

Additional arguments for each measure.

Details

MDA(): Mean Directional Accuracy MDV(): Mean Directional Value MDPV(): Mean Directional Percentage Value

References

Blaskowitz and H. Herwartz (2011) "On economic evaluation of directional forecasts". International Journal of Forecasting, 27(4), 1058-1065.