fabletools: Core Tools for Packages in the 'fable' Framework

Provides tools, helpers and data structures for developing models and time series functions for 'fable' and extension packages. These tools support a consistent and tidy interface for time series modelling and analysis.

See also

Author

Maintainer: Mitchell O'Hara-Wild mail@mitchelloharawild.com (ORCID)

Authors:

Rob Hyndman

Earo Wang (ORCID)

Other contributors:

Di Cook [contributor]

George Athanasopoulos [contributor]

David Holt [contributor]