Create a feature set from tags

Construct a feature set from features available in currently loaded packages. Lists of available features can be found in the following pages:

Features by package

Features by tag

feature_set(pkgs = NULL, tags = NULL)

Arguments

pkgs
The package(s) from which to search for features. If NULL, all registered features from currently loaded packages will be searched.

tags
Tags used to identify similar groups of features. If NULL, all tags will be included.

Registering features

Features can be registered for use with the feature_set() function using register_feature(). This function allows you to register a feature along with the tags associated with it. If the features are being registered from within a package, this feature registration should happen at load time using [.onLoad()].