Features by package

This documentation lists all available in currently loaded packages. This is a useful reference for making a feature_set() from particular package(s).

Details

See the following help topics for more details about currently available features:

feasts

feat_stl

feat_acf

feat_pacf

feat_intermittent

guerrero

unitroot_kpss

unitroot_pp

unitroot_ndiffs

unitroot_nsdiffs

box_pierce

ljung_box

var_tiled_var

var_tiled_mean

shift_level_max

shift_var_max

shift_kl_max

feat_spectral

n_crossing_points

longest_flat_spot

coef_hurst

stat_arch_lm

See also