Forecast plot

Generates forecasts from the given model and adds them to the plot.

GeomForecast

geom_forecast(mapping = NULL, data = NULL, stat = "forecast",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, level = c(80, 95), h = NULL,
  model = fable::ETS(y), fc_args = list(), ...)

StatForecast

Arguments

mapping
Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
The data to be displayed in this layer. There are three options:

If NULL, the default, the data is inherited from the plot data as specified in the call to ggplot().

A data.frame, or other object, will override the plot data. All objects will be fortified to produce a data frame. See fortify() for which variables will be created.

A function will be called with a single argument, the plot data. The return value must be a data.frame, and will be used as the layer data. A function can be created from a formula (e.g. ~ head(.x, 10)).

stat
Use to override the default connection between geom_smooth() and stat_smooth().

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

na.rm
If FALSE, the default, missing values are removed with a warning. If TRUE, missing values are silently removed.

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().

level
A vector of numbers between 0 and 100 which define the confidence range to be plotted. If NULL, confidence intervals will not be plotted, giving only the forecast line.

h
The forecast horison (can be used instead of new_data for regular time series with no exogenous regressors).

model
The time-series model used to produce the forecast. The data must be y (indicating aesthetic y), and the time index for y is determined from the x aesthetic.

fc_args
A list of arguments to be used in the forecast function

...
Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

Format

An object of class GeomForecast (inherits from Geom, ggproto, gg) of length 7.

Value

A layer for a ggplot graph.

Details

The aesthetics required for the forecasting to work includes forecast observations on the y axis, and the time of the observations on the x axis. Refer to the examples below. To automatically set up aesthetics, use autoplot.

See also

Examples

library(ggplot2)
library(tsibble)
as_tsibble(cbind(mdeaths, fdeaths)) %>%
 autoplot() +
 geom_forecast()

#> Plot variable not specified, automatically selected `.vars = value`

#> Error in geom_forecast(): could not find function "geom_forecast"