Glance a mable

Uses the models within a mable to produce a one row summary of their fits. This typically contains information about the residual variance, information criterion, and other relevant summary statistics. Each model will be represented with a row of output.

# S3 method for mdl_df
glance(x, ...)

# S3 method for mdl_ts
glance(x, ...)

Arguments

x
A mable.

...
Arguments for model methods.

Examples

if (requireNamespace("fable", quietly = TRUE)) {
library(fable)
library(tsibbledata)

olympic_running %>%
  model(lm = TSLM(log(Time) ~ trend())) %>%
  glance()
}

#> New names:
#> * glanced -> glanced...1
#> * glanced -> glanced...2
#> * glanced -> glanced...3
#> * glanced -> glanced...4
#> * glanced -> glanced...5
#> * ...

#> # A tibble: 14 x 17
#>    Length Sex   .model r_squared adj_r_squared  sigma2 statistic  p_value    df
#>  *  <int> <chr> <chr>      <dbl>         <dbl>   <dbl>     <dbl>    <dbl> <int>
#>  1    100 men   lm      NA             NA      4.31e-4   NA      NA           2
#>  2    100 women lm      NA             NA      3.84e-4   NA      NA           2
#>  3    200 men   lm      NA             NA      2.47e-4   NA      NA           2
#>  4    200 women lm       0.688          0.669  4.77e-4   35.3     2.07e-5     2
#>  5    400 men   lm      NA             NA      5.10e-4   NA      NA           2
#>  6    400 women lm       0.311          0.253  4.21e-4    5.41    3.83e-2     2
#>  7    800 men   lm      NA             NA      7.93e-4   NA      NA           2
#>  8    800 women lm      NA             NA      7.42e-4   NA      NA           2
#>  9   1500 men   lm      NA             NA      1.12e-3   NA      NA           2
#> 10   1500 women lm       0.158          0.0738 4.40e-4    1.88    2.01e-1     2
#> 11   5000 men   lm      NA             NA      3.58e-4   NA      NA           2
#> 12   5000 women lm       0.00902       -0.239  1.02e-3    0.0364  8.58e-1     2
#> 13  10000 men   lm      NA             NA      2.63e-4   NA      NA           2
#> 14  10000 women lm       0.803          0.770  1.02e-4   24.4     2.60e-3     2
#> # … with 8 more variables: log_lik <dbl>, AIC <dbl>, AICc <dbl>, BIC <dbl>,
#> #   CV <dbl>, deviance <dbl>, df.residual <int>, rank <int>