Compute hilo intervals

Used to extract a specified prediction interval at a particular confidence level from a distribution or fable.

# S3 method for fbl_ts
hilo(x, level = c(80, 95), ...)

hilo(x, ...)

# S3 method for fcdist
hilo(x, level = 95, ...)

Arguments

x
Object to create hilo from

level
The confidence levels for the plotted prediction intervals.

...
Additional arguments for the distribution's quantile function.

Examples

dist_normal(10, 3) %>% hilo(95)

#> [4.120108, 15.87989]95

if (requireNamespace("fable", quietly = TRUE)) {
library(fable)
library(tsibbledata)
library(dplyr)
aus_production %>%
  model(ets = ETS(log(Beer) ~ error("M") + trend("Ad") + season("A"))) %>%
  forecast(h = "3 years") %>%
  mutate(interval = hilo(.distribution, 95))
}

#> # A fable: 12 x 5 [1Q]
#> # Key:     .model [1]
#>    .model Quarter  Beer .distribution                   interval
#>    <chr>    <qtr> <dbl> <dist>                            <hilo>
#>  1 ets    2010 Q3  407. t(N(6.0, 0.0013)) [378.9423, 436.3591]95
#>  2 ets    2010 Q4  483. t(N(6.2, 0.0014)) [448.2540, 519.6750]95
#>  3 ets    2011 Q1  419. t(N(6.0, 0.0014)) [388.4926, 450.8578]95
#>  4 ets    2011 Q2  384. t(N(6.0, 0.0015)) [355.8271, 414.5385]95
#>  5 ets    2011 Q3  405. t(N(6.0, 0.0019)) [371.2483, 440.5715]95
#>  6 ets    2011 Q4  481. t(N(6.2, 0.0022)) [438.2139, 526.1534]95
#>  7 ets    2012 Q1  417. t(N(6.0, 0.0023)) [378.9369, 457.7836]95
#>  8 ets    2012 Q2  383. t(N(5.9, 0.0025)) [346.3522, 422.0245]95
#>  9 ets    2012 Q3  403. t(N(6.0, 0.0032)) [360.6877, 449.6076]95
#> 10 ets    2012 Q4  479. t(N(6.2, 0.0036)) [425.1726, 537.9426]95
#> 11 ets    2013 Q1  416. t(N(6.0, 0.0039)) [367.1605, 468.8992]95
#> 12 ets    2013 Q2  382. t(N(5.9, 0.0043)) [335.2233, 432.9358]95