Unpack a hilo column

This function is superceded. It is recommended that you use the functionality from the distributional package to extract elements from a <hilo> object. For example, you can access the lower bound with <hilo>$lower.

Allows a hilo column to be unpacked into its component columns: "lower", "upper", and "level".

unpack_hilo(data, cols, names_sep = "_", names_repair = "check_unique")

Arguments

data

A data frame.

cols

Name of hilo columns to unpack.

names_sep

If NULL, the default, the names will be left as is. In pack(), inner names will come from the former outer names; in unpack(), the new outer names will come from the inner names.

If a string, the inner and outer names will be used together. In unpack(), the names of the new outer columns will be formed by pasting together the outer and the inner column names, separated by names_sep. In pack(), the new inner names will have the outer names + names_sep automatically stripped. This makes names_sep roughly symmetric between packing and unpacking.

names_repair

Used to check that output data frame has valid names. Must be one of the following options:

"minimal": no name repair or checks, beyond basic existence,

"unique": make sure names are unique and not empty,

"check_unique": (the default), no name repair, but check they are unique,

"universal": make the names unique and syntactic

a function: apply custom name repair.

tidyr_legacy: use the name repair from tidyr 0.8.

a formula: a purrr-style anonymous function (see rlang::as_function())

See vctrs::vec_as_names() for more details on these terms and the strategies used to enforce them.

See also