Features by tag

This documentation lists all available in currently loaded packages. This is a useful reference for making a feature_set() from particular tag(s).

Details

See the following help topics for more details about each feature:
acf

feat_acf

autocorrelation

feat_acf

feat_pacf

boxcox

guerrero

coefficients

coef_hurst

count

n_crossing_points

n_flat_spots

decomposition

feat_stl

intermittent

feat_intermittent

lumpiness

var_tiled_var

optimisation

guerrero

pacf

feat_pacf

portmanteau

box_pierce

ljung_box

rle

n_flat_spots

roll

shift_level_max

shift_var_max

shift_kl_max

seasonal

feat_stl

unitroot_nsdiffs

slide

shift_level_max

shift_var_max

shift_kl_max

spectral

feat_spectral

stability

var_tiled_mean

stl

feat_stl

test

unitroot_kpss

unitroot_pp

unitroot_ndiffs

unitroot_nsdiffs

box_pierce

ljung_box

stat_arch_lm

tile

var_tiled_var

var_tiled_mean

trend

feat_stl

unitroot

unitroot_kpss

unitroot_pp

unitroot_ndiffs

unitroot_nsdiffs

See also