Generate responses from a mable

Use a model's fitted distribution to simulate additional data with similar behaviour to the response. This is a tidy implementation of simulate.

# S3 method for mdl_df
generate(x, new_data = NULL, h = NULL, times = 1,
  seed = NULL, ...)

# S3 method for mdl_ts
generate(x, new_data = NULL, h = NULL, times = 1,
  seed = NULL, ...)

Arguments

x
A mable.

new_data
The data to be generated (time index and exogenous regressors)

h
The simulation horizon (can be used instead of new_data for regular time series with no exogenous regressors).

times
The number of replications.

seed
The seed for the random generation from distributions.

...
Additional arguments for individual simulation methods.

Details

Innovations are sampled by the model's assumed error distribution. If bootstrap is TRUE, innovations will be sampled from the model's residuals. If new_data contains the .innov column, those values will be treated as innovations for the simulated paths..

Examples

library(dplyr)
library(fable)
UKLungDeaths <- as_tsibble(cbind(mdeaths, fdeaths), pivot_longer = FALSE)
UKLungDeaths %>%
  model(lm = TSLM(mdeaths ~ fourier("year", K = 4) + fdeaths)) %>%
  generate(UKLungDeaths, times = 5)

#> # A tsibble: 360 x 5 [1M]
#> # Key:       .model, .rep [5]
#>    new_data           .model  .rep    index  .sim
#>    <list>             <chr>  <int>    <mth> <dbl>
#>  1 <tsibble [72 × 3]> lm         1 1974 Jan 2286.
#>  2 <tsibble [72 × 3]> lm         1 1974 Feb 1779.
#>  3 <tsibble [72 × 3]> lm         1 1974 Mar 2114.
#>  4 <tsibble [72 × 3]> lm         1 1974 Apr 1761.
#>  5 <tsibble [72 × 3]> lm         1 1974 May 1411.
#>  6 <tsibble [72 × 3]> lm         1 1974 Jun 1130.
#>  7 <tsibble [72 × 3]> lm         1 1974 Jul 1230.
#>  8 <tsibble [72 × 3]> lm         1 1974 Aug 1072.
#>  9 <tsibble [72 × 3]> lm         1 1974 Sep 1040.
#> 10 <tsibble [72 × 3]> lm         1 1974 Oct 1543.
#> # … with 350 more rows