Glance a mable

Uses the models within a mable to produce a one row summary of their fits. This typically contains information about the residual variance, information criterion, and other relevant summary statistics. Each model will be represented with a row of output.

# S3 method for class 'mdl_df'
glance(x, ...)

# S3 method for class 'mdl_ts'
glance(x, ...)

Arguments

x

A mable.

...

Arguments for model methods.

Examples

library(fable)
library(tsibbledata)

olympic_running %>%
  model(lm = TSLM(log(Time) ~ trend())) %>% 
  glance()
#> # A tibble: 14 × 17
#>    Length Sex   .model r_squared adj_r_squared   sigma2 statistic  p_value    df
#>     <int> <chr> <chr>      <dbl>         <dbl>    <dbl>     <dbl>    <dbl> <int>
#>  1    100 men   lm       0.829          0.823  0.000431  126.     1.78e-11     2
#>  2    100 women lm       0.754          0.741  0.000384   58.1    3.41e- 7     2
#>  3    200 men   lm       0.886          0.882  0.000247  195.     2.66e-13     2
#>  4    200 women lm       0.688          0.669  0.000477   35.3    2.07e- 5     2
#>  5    400 men   lm       0.843          0.837  0.000510  139.     6.04e-12     2
#>  6    400 women lm       0.311          0.253  0.000421    5.41   3.83e- 2     2
#>  7    800 men   lm       0.774          0.765  0.000793   85.8    1.46e- 9     2
#>  8    800 women lm       0.657          0.633  0.000742   26.8    1.40e- 4     2
#>  9   1500 men   lm       0.703          0.692  0.00112    61.5    2.55e- 8     2
#> 10   1500 women lm       0.158          0.0738 0.000440    1.88   2.01e- 1     2
#> 11   5000 men   lm       0.814          0.806  0.000358   96.6    1.66e- 9     2
#> 12   5000 women lm       0.00902       -0.239  0.00102     0.0364 8.58e- 1     2
#> 13  10000 men   lm       0.902          0.897  0.000263  202.     1.45e-12     2
#> 14  10000 women lm       0.803          0.770  0.000102   24.4    2.60e- 3     2
#> # ℹ 8 more variables: log_lik <dbl>, AIC <dbl>, AICc <dbl>, BIC <dbl>,
#> #   CV <dbl>, deviance <dbl>, df.residual <int>, rank <int>