Unpack a hilo column

Allows a hilo column to be unpacked into its component columns: "lower", "upper", and "level".

unpack_hilo(data, cols, names_sep = "_", names_repair = "check_unique")

Arguments

data
A data frame.

cols
Name of hilo columns to unpack.

names_sep
If NULL, the default, the names of new columns will come directly from the inner data frame.

If a string, the names of the new columns will be formed by pasting together the outer column name with the inner names, separated by names_sep.

names_repair
Used to check that output data frame has valid names. Must be one of the following options:

"minimal": no name repair or checks, beyond basic existence,

"unique": make sure names are unique and not empty,

"check_unique": (the default), no name repair, but check they are unique,

"universal": make the names unique and syntactic

a function: apply custom name repair.

tidyr_legacy: use the name repair from tidyr 0.8.

a formula: a purrr-style anonymous function (see rlang::as_function())

See vctrs::vec_as_names() for more details on these terms and the strategies used to enforce them.

See also